returnToInt

+++ Exercise5_2_Pumin_L.py
@@ -1,5 +1,5 @@
-s = float(input("Input S(mm) :",))
-t = float(input("Iuput Time(s) : "))
+s = int(input("Input S(mm) :",))
+t = int(input("Iuput Time(s) : "))
 print("From v=s/t")
 print("v =",s,"/",t,"=",s/t,"m/s")
